requirements.txt
icrawler>=0.6.10
pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pathlib>=1.0.1
imagehash>=4.3.1
requests>=2.30.0
matplotlib>=3.7.0
selenium>=4.9.0
webdriver-manager>=3.8.6
fake-useragent>=1.1.1
beautifulsoup4==4.8.2  # Downgraded to be compatible with rotating-free-proxies
rotating-free-proxies>=0.1.1
undetected-chromedriver>=3.5.0
python-anticaptcha>=1.0.0
brotli>=1.0.9
# Added dependencies for enhanced functionality
urllib3<2.0.0  # For compatibility with requests
requests-toolbelt>=1.0.0
retry>=0.9.2
python-magic>=0.4.27  # For better file type detection
colorama>=0.4.6  # For better terminal output
psutil>=5.9.5  # For system monitoring
py-cpuinfo>=9.0.0  # For CPU detection
treelib>=1.6.1  # For directory structure visualization
pydantic>=1.10.8  # For data validation
tenacity>=8.2.2  # For advanced retry logic
concurrent-log-handler>=0.9.24  # For better logging
progressbar2>=4.2.0  # Additional progress visualization
# Alternative proxies library that doesn't have the BeautifulSoup conflict
free-proxy>=1.1.0